requirements.txt
2captcha-python==1.0.3
absl-py==0.11.0
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
aioredis==1.3.1
albumentations==1.1.0
amqp==5.0.6
appdirs==1.4.4
argon2-cffi==21.1.0
asgiref==3.4.1
astroid==2.4.2
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.4.0
audioread==2.1.9
backcall==0.2.0
base58==2.1.0
beautifulsoup4==4.9.3
bidict==0.21.4
billiard==3.6.4.0
bleach==4.1.0
blessings==1.7
blis==0.7.4
Bottleneck==1.3.2
Brotli==1.0.9
bs4==0.0.1
cachetools==4.2.0
cardano==0.7.1
catalogue==2.0.6
celery==5.1.2
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
chardet==3.0.4
ci-sdr==0.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorful==0.5.4
ConfigArgParse==1.5.3
ctc-segmentation==1.7.1
currency.converter==0.5.5
CurrencyConverter==0.16.5
cycler==0.11.0
cymem==2.0.5
daphne==3.0.2
debugpy==1.5.1
decorator==4.4.2
discord==1.7.3
discord.py==1.7.3
Distance==0.1.3
distlib==0.3.1
Django==3.2.7
django-allauth==0.49.0
django-captcha==1.0.0
django-cardano==1.28.1
django-channels==0.7.0
django-cors-headers==3.8.0
django-countries==7.2.1
django-crispy-forms==1.12.0
django-eventstream==4.2.0
django-filter==21.1
django-grip==3.0.0
django-node-assets==0.9.9
django-qr-code==2.2.0
django-ranged-response==0.2.0
django-ratelimit==3.0.1
django-rest-auth==0.9.5
django-simple-captcha==0.5.14
djangorestframework==3.13.1
djangorestframework-api-key==2.2.0
dlib==19.24.0
dnspython==2.1.0
espnet-model-zoo==0.1.7
espnet-tts-frontend==0.0.3
fastai==2.5.2
fastbook==0.0.18
fastcore==1.3.26
fastdownload==0.0.5
fastprogress==1.0.0
fastrelease==0.1.12
filelock==3.0.12
filetype==1.0.8
flatbuffers==1.12
fonttools==4.33.3
fsspec==2021.11.0
fvcore==0.1.5.post20220504
gevent==21.8.0
ghapi==0.1.19
google-api-core==1.26.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googleapis-common-protos==1.53.0
gpustat==0.6.0
grakn-client==1.8.1
graphviz==0.17
greenlet==1.1.1
gripcontrol==4.1.0
grpcio==1.46.0
hiredis==2.0.0
huggingface-hub==0.4.0
humanfriendly==10.0
idna==2.8
imageio==2.9.0
imageio-ffmpeg==0.4.7
import-ipynb==0.1.3
importlib-metadata==4.11.3
imutils==0.5.4
iopath==0.1.9
ipaddress==1.0.23
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.5.2
jaconv==0.3
jamo==0.4.1
jedi==0.17.2
Jinja2==2.11.2
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.8.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kaldiio==2.17.2
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
kombu==5.1.0
lazy-object-proxy==1.4.3
levenshtein==0.12.0
librosa==0.9.1
llvmlite==0.38.0
lxml==4.5.2
Markdown==3.3.6
matplotlib==3.5.2
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mnemonic==0.20
monero==0.9.2
moviepy==1.0.3
msgpack==1.0.2
multidict==5.1.0
murmurhash==1.0.5
natsort==8.1.0
nbclient==0.5.4
nbconvert==5.6.1
nbdev==1.1.23
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.8
nltk==3.6.5
notebook==6.4.4
numba==0.55.1
numexpr==2.7.3
numpy==1.21.6
nvidia-ml-py3==7.352.0
oauthlib==3.2.0
onnx==1.8.0
onnx2keras==0.0.24
opencensus==0.7.12
opencensus-context==0.1.2
opencv-python==4.5.5.64
opencv-python-headless==4.5.1.48
opt-einsum==3.3.0
orderedset==2.0.3
packaging==21.3
pandocfilters==1.5.0
parso==0.7.1
pathy==0.6.0
pbkdf2==1.3
pickleshare==0.7.5
Pillow==9.1.0
pipenv==2020.8.13
pooch==1.5.2
portalocker==2.4.0
preshed==3.0.5
proglog==0.1.10
prometheus-client==0.10.0
prompt-toolkit==3.0.8
protobuf==3.14.0
psycopg2==2.9.2
ptyprocess==0.7.0
pubcontrol==3.2.0
py-spy==0.3.5
pyAesCrypt==6.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0
pycparser==2.20
pycryptodomex==3.12.0
pydantic==1.8.2
pyDeprecate==0.3.1
Pygments==2.10.0
pylint==2.6.0
pyparsing==3.0.8
pyrsistent==0.17.3
python-dateutil==2.8.2
python-engineio==4.3.1
python-socketio==5.5.2
python3-openid==3.2.0
pytorch-lightning==1.6.3
pytorch-wpe==0.0.1
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
qudida==0.0.4
random-user-agent==1.0.1
ray==1.2.0
redis==3.5.3
regex==2021.11.10
requests==2.21.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
resampy==0.2.2
Resemblyzer==0.1.1.dev0
rsa==4.7
scikit-image==0.18.1
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.1
secedgar==0.2.1
segno==1.3.3
selenium==3.8.1
Send2Trash==1.8.0
Shapely==1.7.1
six==1.15.0
sklearn==0.0
smart-open==5.2.1
SoundFile==0.10.3.post1
soupsieve==2.0.1
spacy==3.1.3
spacy-legacy==3.0.8
sqlparse==0.4.1
srsly==2.4.1
tabulate==0.8.9
tarsafe==0.0.3
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.7.0
tensorboardX==2.2
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
thinc==8.0.10
threadpoolctl==2.1.0
tifffile==2021.3.17
timm==0.4.5
toml==0.10.1
torch==1.11.0
torch-complex==0.4.3
torch-tb-profiler==0.2.1
torch-vision==0.1.6.dev0
torchaudio==0.11.0
torchmetrics==0.6.0
tornado==6.1
tqdm==4.64.0
traitlets==5.0.5
typeguard==2.13.3
typer==0.4.0
typing==3.7.4.3
typing_extensions==4.2.0
ua-parser==0.10.0
unitypackage-extractor==1.1.0
urllib3==1.24.3
varint==1.0.2
vine==5.0.0
virtualenv==20.0.35
virtualenv-clone==0.5.4
wasabi==0.8.2
wcwidth==0.2.5
webrtcvad==2.0.10
Werkzeug==2.1.2
widgetsnbextension==3.5.1
wrapt==1.12.1
yacs==0.1.8
yarl==1.6.3
zipp==3.8.0
zope.event==4.5.0
